Create ToolbarHelper and move duplicated code
