+++ firefox-ios/Client/Frontend/Browser/Toolbars/ToolbarHelper.swift
@@ -1,0 +1,17 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+
+class ToolbarHelper {
+    func shouldShowNavigationToolbarForTraitCollection(_ traitCollection: UITraitCollection) -> Bool {
+        return traitCollection.verticalSizeClass != .compact
+               && traitCollection.horizontalSizeClass != .regular
+    }
+
+    func shouldShowTopTabsForTraitCollection(_ traitCollection: UITraitCollection) -> Bool {
+        return traitCollection.verticalSizeClass == .regular
+               && traitCollection.horizontalSizeClass == .regular
+    }
+}
